+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file.
 * Unmerging suites no longer fails if the merged suite had the same name as one of its constituent suites ([#790] https://gitlab.com/yale-sdmp/vesta/issues/790).
 * Fix updating groups through the superuser dashboard ([#775](https://gitlab.com/yale-sdmp/vesta/issues/775)).
 * Bulk user enrollment no longer fails silently when users exist and now flashes an error message ([#156](https://gitlab.com/yale-sdmp/vesta/issues/156)).
+* Admins can add users to groups even if users have pending invitations ([#773](https://gitlab.com/yale-sdmp/vesta/issues/773)).
 
 ## v1.0.8 - 2018-04-09
 ### Fixed